medicine table is ready
